create folders to better separate and structure the files of each component and hook
