+++ components/content/index.ts
@@ -1,0 +1,5 @@
+export * from './CurrentFavTech';
+export * from './AboutMe';
+export * from './FavProjects';
+export * from './TimeLineExp';
+export * from './ProcessWork';
